requirements.txt
fastapi
uvicorn
pydantic
httpx
tenacity
python-dotenv
